Add display.js, count-comment.js files
