+++ src/count-comment.js
@@ -1,0 +1,6 @@
+const commentCounter = (json) => {
+  const { length } = json;
+  document.querySelector('.comment-heading').innerHTML = `Comments(${length})`;
+};
+
+export default commentCounter;
